+++ src/PyF/Internal/PythonSyntax.hs
@@ -9,8 +9,7 @@
 This module provides a parser for <https://docs.python.org/3.4/library/string.html#formatspec python format string mini language>.
 -}
 module PyF.Internal.PythonSyntax
-  ( parsePythonFormatString
-  , parseGenericFormatString
+  ( parseGenericFormatString
   , Item(..)
   , FormatMode(..)
   , Padding(..)
@@ -19,6 +18,8 @@ module PyF.Internal.PythonSyntax
   , AlternateForm(..)
   , pattern DefaultFormatMode
   , Parser
+  , ParsingContext(..)
+  , ExprOrValue(..)
   )
 where
 
@@ -28,6 +29,7 @@ import Text.Megaparsec
 import qualified Text.Megaparsec.Char.Lexer as L
 import Text.Megaparsec.Char
 import Data.Void (Void)
+import Control.Monad.Reader
 
 import qualified Data.Char
 
@@ -40,7 +42,13 @@ import qualified Language.Haskell.Exts.Extension as ParseExtension
 import qualified Language.Haskell.Exts.SrcLoc as SrcLoc
 import PyF.Formatters
 
-type Parser t = Parsec Void String t
+type Parser t = ParsecT Void String (Reader ParsingContext) t
+
+data ParsingContext = ParsingContext
+  { delimiters :: (Char, Char)
+  , enabledExtensions :: [ParseExtension.Extension]
+  }
+  deriving (Show)
 
 {-
 -- TODO:
@@ -82,14 +90,14 @@ Right [
                       (FixedF (Precision 2) NormalForm Minus)
                        Nothing))]
 -}
-parsePythonFormatString :: [ParseExtension.Extension] -> Parser [Item]
-parsePythonFormatString exts = parseGenericFormatString exts ('{', '}')
 
-parseGenericFormatString :: [ParseExtension.Extension] -> (Char, Char) -> Parser [Item]
-parseGenericFormatString exts delimiters = many (rawString delimiters <|> escapedParenthesis delimiters <|> replacementField exts delimiters) <* eof
+parseGenericFormatString :: Parser [Item]
+parseGenericFormatString = do
+  many (rawString <|> escapedParenthesis <|> replacementField) <* eof
 
-rawString :: (Char, Char) -> Parser Item
-rawString (openingChar,closingChar) = do
+rawString :: Parser Item
+rawString = do
+  (openingChar, closingChar) <- delimiters <$> ask
   chars <- some (noneOf ([openingChar, closingChar]))
   case escapeChars chars of
     Left remaining -> do
@@ -98,8 +106,11 @@ rawString (openingChar,closingChar) = do
       fancyFailure (Set.singleton (ErrorFail "lexical error in literal section"))
     Right escaped -> return (Raw escaped)
 
-escapedParenthesis :: (Char, Char) -> Parser Item
-escapedParenthesis (openingChar, closingChar) = Raw <$> (parseRaw openingChar <|> parseRaw closingChar)
+escapedParenthesis :: Parser Item
+escapedParenthesis = do
+  (openingChar, closingChar) <- delimiters <$> ask
+
+  Raw <$> (parseRaw openingChar <|> parseRaw closingChar)
   where parseRaw c = c:[] <$ string (replicate 2 c)
 
 {- | Replace escape chars with their value. Results in a Left with the
@@ -118,8 +129,11 @@ escapeChars s = case Data.Char.readLitChar s of
                   ((c, xs):_) -> (c :) <$> escapeChars xs
                   _ -> Left s
 
-replacementField :: [ParseExtension.Extension] -> (Char, Char) -> Parser Item
-replacementField exts (charOpening, charClosing) = do
+replacementField :: Parser Item
+replacementField = do
+  exts <- enabledExtensions <$> ask
+  (charOpening, charClosing) <- delimiters <$> ask
+
   _ <- char charOpening
   expr <- evalExpr exts (many (noneOf (charClosing:":" :: [Char])))
   fmt <- optional $ do
@@ -142,9 +156,15 @@ data Padding = PaddingDefault
              | Padding Integer (Maybe (Maybe Char, AnyAlign))
              deriving (Show)
 
+-- | Represents a value of type @t@ or an Haskell expression supposed to represents that value
+data ExprOrValue t
+  = Value t
+  | HaskellExpr Exp
+  deriving (Show)
+
 -- | Floating point precision
 data Precision = PrecisionDefault
-               | Precision Integer
+               | Precision (ExprOrValue Integer)
                deriving (Show)
 {-
 
@@ -231,7 +251,7 @@ format_spec = do
 
   grouping <- optional grouping_option
 
-  prec <- option PrecisionDefault (char '.' *> (Precision <$> precision))
+  prec <- option PrecisionDefault parsePrecision
   t <- optional type_
 
   let padding = case w of
@@ -244,6 +264,17 @@ format_spec = do
       Right fmt -> pure (FormatMode padding fmt grouping)
       Left typeError -> do
         lastCharFailed typeError
+
+parsePrecision :: Parser Precision
+parsePrecision = do
+  exts <- enabledExtensions <$> ask
+  (charOpening, charClosing) <- delimiters <$> ask
+
+  _ <- char '.'
+  choice [
+    Precision . Value <$> precision,
+    char charOpening *> (Precision . HaskellExpr <$> evalExpr exts (manyTill anySingle (char charClosing)))
+    ]
 
 evalFlag :: TypeFlag -> Padding -> Maybe Char -> Precision -> AlternateForm -> Maybe SignMode -> Either String TypeFormat
 evalFlag Flagb _pad _grouping prec alt s = failIfPrec prec (BinaryF alt (defSign s))
@@ -279,7 +310,11 @@ errgGn = "Use one of {'b', 'c', 'd', 'e', 'E', 'f', 'F', 'g', 'G', 'n', 'o', 's'
 
 failIfPrec :: Precision -> TypeFormat -> Either String TypeFormat
 failIfPrec PrecisionDefault i = Right i
-failIfPrec (Precision i) _ = Left ("Type incompatible with precision (." ++ show i ++ "), use any of {'e', 'E', 'f', 'F', 'g', 'G', 'n', 's', '%'} or remove the precision field.")
+failIfPrec (Precision e) _ = Left ("Type incompatible with precision (." ++ showExpr ++ "), use any of {'e', 'E', 'f', 'F', 'g', 'G', 'n', 's', '%'} or remove the precision field.")
+  where
+    showExpr = case e of
+      Value v -> show v
+      HaskellExpr expr -> show expr
 
 failIfAlt :: AlternateForm -> TypeFormat -> Either String TypeFormat
 failIfAlt NormalForm i = Right i
@@ -353,8 +388,3 @@ type_ = choice [
   FlagX <$ char 'X',
   FlagPercent <$ char '%'
   ]
-
-
-  -- TODO: remove !
-deriving instance Lift Precision
-deriving instance Lift Padding